VIRO-4599: compile protobuf using libc++
For compatibility with NDK r18, create a new
system for compiling protobuf so that it works
with libc++ (insead of GNU STL).

The previous method for compiling protobuf
by using autoconf would not work with libc++
for two reasons:

1. When trying to build a shared libprotobuf-lite.so,
the generated library would always use shared
libc++, where we need static libc++. Trying
numerous different flags to prevent this did not
work because something in the autoconf system
was inserting a -lstdc++ flag in the linker, making
it use the shared lib regardless of any other
setting.

2. If we switched to using the static libprotobuf-lite,
we could not get it to compile properly with -fPIC,
which made the resultant lib incompatible with
ViroRenderer (RELOC_32 errors would result).

To fix this, this commit introduce a new method
for generating libprotobuf-lite.a, by using a
shell Android application that exists only to
compile protobuf-lite from source. This
guarantees we use a toolchain that’s compatible
with ViroRenderer.

For building and updating details, see Method
2 in the updated README.